keep pollers under internal/pollers, fix error message
