+++ sdk/azcore/internal/pollers/armloc/loc.go
@@ -4,7 +4,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-package loc
+package armloc
 
 import (
 	"context"
